src/requirement.txt
pact-python == 1.4.6
fastapi == 0.72.0
uvicorn == 0.17.0
